Enable empty confirmationPrompt and rejectionStatement.

+++ playground/delete_log_groups.py
@@ -5,8 +5,9 @@ def delete_log_groups():
     client = boto3.client('logs')
     response = client.describe_log_groups()
     for l in response['logGroups']:
-        print(l)
-        client.delete_log_group(logGroupName=l['logGroupName'])
+        if "LexBuilderCustomReseo" in l['logGroupName']:
+            print(l)
+            client.delete_log_group(logGroupName=l['logGroupName'])
 
 
 if __name__ == '__main__':
